% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_data.R
\docType{data}
\name{bg_data}
\alias{bg_data}
\title{Background variables for plausible values estimation}
\format{
A data frame with 5330 rows and 4 variables:
\describe{
  \item{ID_t}{target IDs}
  \item{age}{age of the subjects}
  \item{gender}{gender of the subjects, 1 female}
  \item{math}{math competence of the subjects}
}
}
\usage{
bg_data
}
\description{
The data set contains simulated variables age, gender, math
competence (loosely based on NEPS SC6)
}
\keyword{datasets}
