% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_data.R
\docType{data}
\name{bgdata_sc3}
\alias{bgdata_sc3}
\title{Background variables for plausible values estimation (school cohort example)}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
  \item{ID_t}{target IDs}
  \item{grade}{grade in German class; from 1 (very good) to 6 (very poor/fail)}
  \item{gender}{gender of the subjects, 1 male, 2 female}
  \item{nbooks}{number of books at home; 1 at most, 2 more than 100}
  \item{migration}{migration background; 1 no, 2 yes}
}
}
\usage{
bgdata_sc3
}
\description{
The data set contains simulated variables German grade, gender, number of
books at home and migration (loosely based on NEPS SC3)
}
\keyword{datasets}
