% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_data.R
\docType{data}
\name{bgdata_sc6}
\alias{bgdata_sc6}
\title{Background variables for plausible values estimation (adult cohort example)}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
  \item{ID_t}{target IDs}
  \item{age}{age of the subjects; split in younger (2) and older (1)}
  \item{gender}{gender of the subjects, 1 male, 2 female}
  \item{nbooks}{number of books at home; 1 at most, 2 more than 100}
  \item{migration}{migration background; 1 no, 2 yes}
}
}
\usage{
bgdata_sc6
}
\description{
The data set contains simulated variables age, gender, number of books at
home and migration (loosely based on NEPS SC6)
}
\keyword{datasets}
