% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CART.R
\name{CART}
\alias{CART}
\title{Sequential Classification and Regression Trees
adapted from package LaRA by J.-C. Gaasch (2017)}
\usage{
CART(X, itermcmc, burnin, nmi = 10, thin = 1, cartctrl1 = 5,
  cartctrl2 = 1e-04)
}
\arguments{
\item{X}{data frame containing person-level covariates. They can be numeric or factor
variables and contain missing values coded as \code{NA}. Rows of \code{X} correspond to
persons and columns correspond to covariates. \code{X} has to contain a person
identifier called \code{ID_t}.}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}

\item{nmi}{number of multiple imputations; defaults to 10.}

\item{thin}{thinning interval, i.e., retain only every \code{thin}th iteration (if argument
\code{thin} is used, \code{itermcmc*thin} and \code{burnin*thin} yield the total number of
MCMC and burnin iterations).}

\item{cartctrl1}{minimum number of observations in any terminal CART node during covariates
imputation cycles.}

\item{cartctrl2}{complexity parameter. Any CART split that does not decrease the overall
lack of fit by a factor of \code{control2} is not attempted during covariates imputation
cycles.}
}
\value{
list with \code{nmi} imputed data sets.
}
\description{
Multiply imputes missing data using the CART algorithm.
}
\details{
Partially observed variables are imputed in each
sampling iteration. Sequential CART (Burgette & Reiter, 2010) are utilized as approximations
to the full conditional distributions of missing values in \code{X}.
}
\references{
Burgette, L. F., & Reiter, J. P. (2010). Multiple imputation for missing data
via sequential regression trees. \emph{American Journal of Epidemiology}, \emph{172}(9),
1070-1076.
}
