% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_data.R
\docType{data}
\name{CohortProfile_sc3}
\alias{CohortProfile_sc3}
\title{School IDs for plausible values estimation (school cohort example)}
\format{
A data frame with 12000 rows and 3 variables:
\describe{
  \item{ID_t}{target IDs}
  \item{ID_i}{school IDs}
  \item{wave}{assessment wave}
}
}
\usage{
CohortProfile_sc3
}
\description{
The data set contains simulated variables for school IDs and student - school
mapping
}
\keyword{datasets}
