% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currently_implemented.R
\name{currently_implemented}
\alias{currently_implemented}
\title{Get information about implementation progress of
\code{NEPSscaling::plausible_values()}}
\usage{
currently_implemented()
}
\description{
Get information about implementation progress of
\code{NEPSscaling::plausible_values()}
}
