% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviations_of_package_from_suf.R
\name{deviations_of_package_from_suf}
\alias{deviations_of_package_from_suf}
\title{Get information about known deviations of the NEPSscaling WLE mean
estimation from SUF WLE mean estimation. The competence estimators of SUF
and package correlate to at least 0.95}
\usage{
deviations_of_package_from_suf()
}
\description{
Get information about known deviations of the NEPSscaling WLE mean
estimation from SUF WLE mean estimation. The competence estimators of SUF
and package correlate to at least 0.95
}
