% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tree.R
\name{display_imputation_tree}
\alias{display_imputation_tree}
\title{Display the tree structure returned by the CART algorithm built to impute
one missing variable}
\usage{
display_imputation_tree(pv_obj, imputation, variable)
}
\arguments{
\item{pv_obj}{An object of class "pv_obj" containing at least the sublist
"treeplot"}

\item{imputation}{Integer or character string value. If it is supplied as an
integer it signifies the index of the imputation (ranging from 1 to max.
control$ML$nmi) to choose the imputed variable from. As a character string it
supplies the name of the imputation (e.g., "imp2").}

\item{variable}{A character string. Name of the variable to be displayed.}
}
\value{
a ggplot2 plot object representing the tree structure
}
\description{
Display the tree structure returned by the CART algorithm built to impute
one missing variable
}
