% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tree.R
\name{display_tree}
\alias{display_tree}
\title{Display the tree structure returned by the CART algorithm built to impute
one missing variable}
\usage{
display_tree(tree)
}
\arguments{
\item{tree}{string representation of the tree}
}
\value{
a ggplot2 plot object representing the tree structure
}
\description{
Display the tree structure returned by the CART algorithm built to impute
one missing variable
}
