% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_variable_importance.R
\name{display_variable_importance}
\alias{display_variable_importance}
\title{Display variable importance for specific imputation of a variable}
\usage{
display_variable_importance(pv_obj, imputation, variable)
}
\arguments{
\item{pv_obj}{pv_obj by plausible_values()}

\item{imputation}{name or index of imputation (e.g., 1, or "imp3")}

\item{variable}{specific imputed variable for which the predictor variable
importance shall be displayed}
}
\value{
barplot of feature importance for imputed variable
}
\description{
Display variable importance for specific imputation of a variable
}
