% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_domain}
\alias{get_domain}
\title{Get the domain for which plausible values have been estimated}
\usage{
get_domain(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the competence domain as a String
}
\description{
Get the domain for which plausible values have been estimated
}
