% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_eap}
\alias{get_eap}
\title{Get the estimated EAP values and standard errors for each person}
\usage{
get_eap(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
a data.frame containing the EAPs and their respective standard errors
}
\description{
Get the estimated EAP values and standard errors for each person
}
