% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_eap_reliability}
\alias{get_eap_reliability}
\title{Get the EAP reliability of the plausible values}
\usage{
get_eap_reliability(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the EAP reliability of the plausible values as a numeric value
}
\description{
Get the EAP reliability of the plausible values
}
