% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_exclude}
\alias{get_exclude}
\title{Get exclude}
\usage{
get_exclude(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
a list of matrices or a matrix of AIC and BIC
}
\description{
In the longitudinal case, it might not be wanted that all background variables
are used in the background model of the plausible values for all time points.
The vector or list exclude contains which variables are NOT used in the
background model of the given wave.
}
