% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_imputation_names}
\alias{get_imputation_names}
\title{Get names of used imputations.}
\usage{
get_imputation_names(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
a character vector containing the names of the used imputations
}
\description{
Imputations and the plausible values estimated with them are chosen randomly,
so not all of them might be returned after the estimation. This function
shows which ones were actually retained.
}
