% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_imputation_tree}
\alias{get_imputation_tree}
\title{Get the tree representation of specific imputation.}
\usage{
get_imputation_tree(pv_obj, imputation, variable)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}

\item{imputation}{Integer or character string value. If it is supplied as an
integer it signifies the index of the imputation (ranging from 1 to max.
control$ML$nmi) to choose the imputed variable from. As a character string it
supplies the name of the imputation (e.g., "imp2").}

\item{variable}{A character string. Name of the variable to be displayed.}
}
\value{
a character vector containing the tree structure
}
\description{
For each imputed variable, a CART was constructed. This function returns its
character representation.
}
