% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_include_nr}
\alias{get_include_nr}
\title{Get include_nr}
\usage{
get_include_nr(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
a boolean indicating whether the number of not-reached items as a
proxy for processing speed should be included in background model
}
\description{
Get include_nr
}
