% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_info_criteria}
\alias{get_info_criteria}
\title{Get the information criteria}
\usage{
get_info_criteria(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
a list of matrices or a matrix of AIC and BIC
}
\description{
Get the information criteria
}
