% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_item_difficulties}
\alias{get_item_difficulties}
\title{Get the fixed item difficulties and their standard errors}
\usage{
get_item_difficulties(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the matrix containing the fixed item difficulties and their estimated
standard errors as numeric values
}
\description{
Get the fixed item difficulties and their standard errors
}
