% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_model}
\alias{get_model}
\title{Get the estimation model}
\usage{
get_model(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the estimation model as a String
}
\description{
Get the estimation model
}
