% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_n_testtakers}
\alias{get_n_testtakers}
\title{Get the number of test takers}
\usage{
get_n_testtakers(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
a string of the file path leading to the competence data
}
\description{
Get the number of test takers
}
