% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_nvalid}
\alias{get_nvalid}
\title{Get the number of valid responses (i.e., non-missing) required for a test
taker to be considered during the estimation}
\usage{
get_nvalid(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the minimum number of valid responses as an integer value
}
\description{
Get the number of valid responses (i.e., non-missing) required for a test
taker to be considered during the estimation
}
