% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_nvalid_per_person}
\alias{get_nvalid_per_person}
\title{Get the number of valid responses (i.e., non-missing) for each test taker}
\usage{
get_nvalid_per_person(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
a data.frame containing the number of valid responses per person
as an integer value
}
\description{
Get the number of valid responses (i.e., non-missing) for each test taker
}
