% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_path}
\alias{get_path}
\title{Get the file path}
\usage{
get_path(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
a string of the file path leading to the competence data
}
\description{
Get the file path
}
