% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_posterior_means}
\alias{get_posterior_means}
\title{Get the posterior means of the EAPs, WLEs, and plausible values}
\usage{
get_posterior_means(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the overall mean of the plausible values as a numeric value
}
\description{
Get the posterior means of the EAPs, WLEs, and plausible values
}
