% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_pv_index}
\alias{get_pv_index}
\title{Get the i-th plausible value and its respective background data}
\usage{
get_pv_index(pv.obj, index = 1)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}

\item{index}{the index for the plausible values ranging from 1 to npv
(defaults to 1)}
}
\value{
a data.frame containing the i-th plausible values and its respective
background data
}
\description{
Get the i-th plausible value and its respective background data
}
