% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_pv_list}
\alias{get_pv_list}
\title{Get the complete list of estimated plausible values and their respective
background data}
\usage{
get_pv_list(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the list of npv plausible values and their respective background data
}
\description{
Get the complete list of estimated plausible values and their respective
background data
}
