% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_regression_coefficients}
\alias{get_regression_coefficients}
\title{Get the regression coefficients estimated in the latent regression on the
background data}
\usage{
get_regression_coefficients(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the matrix containing the regression weight for each background
variable, including an intercept, as numeric values
}
\description{
Get the regression coefficients estimated in the latent regression on the
background data
}
