% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_rotation}
\alias{get_rotation}
\title{Get whether the test rotation has been considered during estimation}
\usage{
get_rotation(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
whether the test rotation has been considered as a String
}
\description{
Get whether the test rotation has been considered during estimation
}
