% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contextual_info.R
\name{get_school_id_data}
\alias{get_school_id_data}
\title{import CohortProfile which contains the school ID_i and convert it to wide 
data format}
\usage{
get_school_id_data(path)
}
\description{
import CohortProfile which contains the school ID_i and convert it to wide 
data format
}
