% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_starting_cohort}
\alias{get_starting_cohort}
\title{Get the starting cohort for which plausible values have been estimated}
\usage{
get_starting_cohort(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
the number of the starting cohort as an integer value (see
https://www.neps-data.de/en-us/datacenter/dataanddocumentation.aspx for their
meaning)
}
\description{
Get the starting cohort for which plausible values have been estimated
}
