% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_test_position}
\alias{get_test_position}
\title{Get the position (first or second) the test has been administered in}
\usage{
get_test_position(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
a data.frame containing the position (first or second) the test has
been administered in
}
\description{
Get the position (first or second) the test has been administered in
}
