% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_type}
\alias{get_type}
\title{Get the type of estimation (longitudinal or cross-sectional)}
\usage{
get_type(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
the type of estimation as a String
}
\description{
Get the type of estimation (longitudinal or cross-sectional)
}
