% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_wave}
\alias{get_wave}
\title{Get the wave for which plausible values have been estimated}
\usage{
get_wave(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the number of the wave as an integer value
}
\description{
Get the wave for which plausible values have been estimated
}
