% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_wle_reliability}
\alias{get_wle_reliability}
\title{Get the WLE reliability of the plausible values}
\usage{
get_wle_reliability(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
the WLE reliability of the plausible values as a numeric value
}
\description{
Get the WLE reliability of the plausible values
}
