% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contextual_info.R
\name{get_wle_vnames}
\alias{get_wle_vnames}
\title{get wle variable names for the current SC/domain/waves}
\usage{
get_wle_vnames(waves, SC, domain)
}
\description{
get wle variable names for the current SC/domain/waves
}
