% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEPSscalingGUI.R
\name{NEPSscalingGUI}
\alias{NEPSscalingGUI}
\title{Shiny interface for NEPSscaling}
\usage{
NEPSscalingGUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to \code{\link[shiny]{runApp}}}
}
\description{
This function calls a shiny interface for NEPSscaling.
}
