% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEPSshiny.R
\name{NEPSshiny}
\alias{NEPSshiny}
\title{Shiny interface for NEPSscaling}
\usage{
NEPSshiny(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to \code{\link[shiny]{runApp}}}
}
\description{
This function calls a shiny interface for NEPSscaling.
}
