% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values.R
\name{plausible_values}
\alias{plausible_values}
\title{Function for estimating plausible values with NEPS competence data and NEPS
scaled item parameters}
\usage{
plausible_values(SC, domain = c("MA", "RE", "SC", "IC", "LI", "EF", "NR",
  "NT", "OR", "ST", "BA", "CD", "GR"), wave, method = c("IND", "MMI", "CART"),
  path, filetype = c("SPSS", "Stata"), bgdata = NULL, npv = 10L,
  longitudinal = FALSE, rotation = TRUE, nvalid = 3L, control = list(EAP
  = FALSE, WLE = FALSE, IND = list(varex = 0.9, pca.data = FALSE, vars = NULL),
  MMI = list(nmi = 10L, method = vector("character", length = ncol(bgdata) - 1),
  predictorMatrix = (1 - diag(1, ncol(bgdata) - 1)), visitSequence = NULL, form
  = vector("character", length = ncol(bgdata) - 1), post = vector("character",
  length = ncol(bgdata) - 1), defaultMethod = c("pmm", "logreg", "polyreg",
  "polr"), maxit = 10, diagnostics = TRUE, printFlag = FALSE, seed = NA,
  imputationMethod = NULL, defaultImputationMethod = NULL,      data.init =
  NULL), CART = list(itermcmc = 10000, burnin = 2000, thin = 1, tdf = 10,
  cartctrl1 = 5, cartctrl2 = 1e-04), TAM = list(ntheta = 2000, normal.approx =
  FALSE, samp.regr = FALSE, theta.model = FALSE, np.adj = 8, na.grid = 5)))
}
\arguments{
\item{SC}{numeric. The starting cohort used for the analysis is
indicated by an integer value (e.g., starting cohort one = 1).}

\item{domain}{string. The competence domain of interest is indicated
by the domain abbreviation as used in the variable names (see Fuß, D., Gnambs, 
T., Lockl, K., & Attig, M., 2016).}

\item{wave}{numeric. The wave of competence testing is indicated by
an integer value (e.g., wave one = 1).}

\item{method}{string. The method used for missing data handling in covariates
(IND: missing indicators, MMI: multiple multiple imputation, CART: sequential
classification and regression trees).}

\item{path}{file path leading to the location of the competence data}

\item{filetype}{filetype of competence data (the NEPS RDC provides SPSS
and Stata data files)}

\item{bgdata}{data frame containing background variables. Categorical 
variables have to be specified as factors. If \code{bgdata = NULL}, plausible 
values are estimated without a background model.}

\item{npv}{number of plausible values to be estimated; it defaults
to 10.}

\item{longitudinal}{logical. TRUE indicating that a model based on linked
item difficulties (without correction for position effects) is to be
estimated. Defaults to FALSE.}

\item{rotation}{logical. TRUE indicating that the competence scores
should be corrected for the rotation design of the booklets. Defaults to
TRUE. If both longitudinal and rotation are TRUE, rotation is set to FALSE
automatically.}

\item{nvalid}{minimum number of responses test takers had to give for
inclusion into the estimation process. NEPS default is at least three.}

\item{control}{list of additional options. If \code{EAP = TRUE} or
\code{WLE = TRUE}, the EAPs or WLEs will be returned as well. Furthermore,
additional control options for IND are \code{varex} (the amount of
variance explained by the retained PCA factors), \code{pca.data} (a logical;
if TRUE, no plausible  values are estimated, but the data to estimate them
is returned) and \code{vars} (a vector of variable names that are not to be
included in the PCA). Additional control options for MMI consist of \code{nmi}
(an integer value specifying the number of multiple imputations; 10 by default)
and further options for mice::mice(). Analogously, CART options and TAM options
(used for plausible values estimation in methods IND and MMI) can be specified.
For further information see the documentations of LaRA::mglrm() for CART and
TAM::tam.pv() for TAM.}
}
\value{
\code{plausible_values()} returns an object of class \code{pv.obj}
containing \code{npv} plausible values with the data set used
for estimating them as well as WLEs and EAPs and their respective standard
error for the specified domain, wave and starting cohort of the NEPS.
}
\description{
Function for estimating plausible values with NEPS competence data and NEPS
scaled item parameters
}
\note{
For an extensive example, see the accompanying NEPS Survey Paper.
The function will only work with NEPS data. To access NEPS data see
https://www.neps-data.de/en-us/datacenter/dataaccess.aspx.
}
\examples{
\dontrun{
rm(list = ls())
library(NEPStools)
library(foreign)
## read in data object for conditioning variables
data(bg_data)
bg_data$gender <- as.factor(bg_data$gender)
## define path to NEPS competence data
path <- 'Usr/NEPS-data/'
setwd(path)
## save simulated target competencies to path
data(xTargetCompetencies_sim)
write.dta(dataframe = xTargetCompetencies_sim, file = 'SC6_xTargetCompetencies_sim.dta')
## estimate default number of 10 plausible values
## note: the example background data is completely observed!
result <- plausible_values(SC = 'SC6'
                           , domain = 'RE'
                           , wave = 'w3'
                           , bgdata = bg_data
                           , path = path
                           , method = 'IND'
                           , filetype = 'Stata'
                           )
}

}
\references{
Albert, J. H. (1992). Bayesian estimation of normal ogive item
response curves using Gibbs sampling. \emph{Journal of Educational
Statistics}, \emph{17}(3), 251-269.

Azur, M. J., Stuart, E. A., Frangakis, C., & Leaf, P. J. (2011).
Multiple imputation by chained equations: what is it and how does it work?
\emph{International Journal of Methods in Psychiatric Research, 20}(1), 40-49.

Blossfeld, H.-P., Rossbach, H.-G., & von Maurice,  J. (Eds.) (2011).
Education as a Lifelong Process - The German National Educational Panel Study
(NEPS). \emph{Zeitschrift fuer Erziehungswissenschaft, Sonderheft 14}.

Burgette, L. F., & Reiter, J. P. (2010). Multiple imputation for
missing data via sequential regression trees. \emph{American Journal of
Epidemiology}, \emph{172}(9), 1070-1076.

Mislevy, R. J. (1991). Randomization-based inference about latent
variables from complex samples. \emph{Psychometrika, 56}(2), 177-196.

Pohl, S., & Carstensen, C. H. (2012). NEPS technical report -
Scaling the data of the competence tests.

Scharl, A., Carstensen, C. H., & Gnambs, T. (2018).
\emph{Estimating Plausible Values with NEPS Data: An Example Using Reading
Competence in Starting Cohort 6 (NEPS Survey Paper No. XX)}. Bamberg: Leibniz
Institute for Educational Trajectories, National Educational Panel Study.

Tanner, M. A., & Wong, W. H. (1987). The calculation of posterior
distributions by data augmentation. \emph{Journal of the American Statistical
Association}, \emph{82}(398), 528-549.

Weirich, S., Haag, N., Hecht, M., Boehme, K., Siegle, T., & Luedtke,
O. (2014). Nested multiple imputation in large-scale assessments.
\emph{Large-Scale Assessments in Education, 2}(1), 9.
}
