% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_mdlrm.R
\name{plausible_values_mdlrm}
\alias{plausible_values_mdlrm}
\title{Plausible values imputation for longitudinal data}
\usage{
plausible_values_mdlrm(Y, X = NULL, npv = 10, est.alpha = TRUE,
  itermcmc, burnin, thin = 1, tdf = 10, cartctrl1 = 5,
  cartctrl2 = 1e-04, SC, domain, waves)
}
\arguments{
\item{Y}{test responses}

\item{X}{covariates}

\item{npv}{number of plausible values to draw for each respondent.}

\item{est.alpha}{logical, should discrimination parameters be estimated or
fixed to 1 for dichotomous and to 0.5 for ordinal items}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}

\item{thin}{numeric, thinning interval}

\item{tdf}{degrees of freedom of multivariate-t proposal distribution for category cutoff
parameters for ordinal items.}

\item{cartctrl1}{minimum number of observations in any terminal CART node during covariates
imputation cycles.}

\item{cartctrl2}{complexity parameter. Any CART split that does not decrease the overall
lack of fit by a factor of \code{cartctrl2} is not attempted during covariates imputation
cycles.}

\item{SC}{string, NEPS starting cohort}

\item{domain}{string, competence domain}

\item{waves}{vector of strings, waves of the competence testings}
}
\value{
list with \code{npv} elements, each containing a data frame of the ID, plausible values for
each dimension and, if specified, imputed versions of \code{X}.
}
\description{
This function estimates a multidimensional item response model (nominal and
ordinal data). Missing values in the response data are ignored, CART is used
to impute missing values in the covariate data.
}
