% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pv_obj.R
\name{print.pv_obj}
\alias{print.pv_obj}
\title{Print information about NEPSscaling object}
\usage{
\method{print}{pv_obj}(x, ...)
}
\arguments{
\item{x}{return object of function \code{NEPScaling::plausible_values()}}

\item{...}{unused}
}
\description{
Print information about NEPSscaling object
}
