% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_longitudinal_sc6_reading.R
\name{select_longitudinal_sc6_reading}
\alias{select_longitudinal_sc6_reading}
\title{Because of resampling in wave 5, two distinct samples have the same items}
\usage{
select_longitudinal_sc6_reading(data, SC, domain, min_valid)
}
\arguments{
\item{data}{...}

\item{SC}{..}

\item{domain}{...}

\item{min_valid}{...}
}
\description{
Because of resampling in wave 5, two distinct samples have the same items
}
