% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pv_obj.R
\name{summary.pv_obj}
\alias{summary.pv_obj}
\title{Print information about NEPSscaling plausible values estimation}
\usage{
\method{summary}{pv_obj}(object, ...)
}
\arguments{
\item{object}{return object of function \code{NEPScaling::plausible_values()}}

\item{...}{unused}
}
\description{
Print information about NEPSscaling plausible values estimation
}
