% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pv.R
\name{write_pv}
\alias{write_pv}
\title{Function to write pv.obj to file}
\usage{
write_pv(pv.obj, path, ext = c("SPSS", "Stata", "Mplus"))
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}

\item{path}{path to the directory where the plausible values are going to
be saved to}

\item{ext}{format (either SPSS, Stata (version 14) or Mplus compatible)}
}
\description{
Function to write pv.obj to file
}
\examples{
\dontrun{
rm(list = ls())
library(NEPScaling)
# the pv.obj results was created by plausible_values()
path <- "chosen/directory/"
write_pv(results, path, "Mplus")
write_pv(results, path, "Stata")
write_pv(results, path, "SPSS")
}

}
