% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xTargetCompetencies_sim.R
\docType{data}
\name{xTargetCompetencies_sim}
\alias{xTargetCompetencies_sim}
\title{Competence items for plausible values estimation}
\format{A data frame with 5330 rows and 30 variables:
\describe{
  \item{ID_t}{target IDs}
  \item{wave_w3}{competence test indicator for wave 3}
  \item{tx80211_w3}{booklet rotation indicator}
  \item{rea30110_c}{item of reading competence test, 0 incorrect answer, 1 correct answer}
  \item{rea3012s_c}{...}
  \item{rea30130_c}{...}
  \item{rea30140_c}{...}
  \item{rea3015s_c}{...}
  \item{rea30210_c}{...}
  \item{rea30220_c}{...}
  \item{rea30230_c}{...}
  \item{rea30240_c}{...}
  \item{rea30250_c}{...}
  \item{rea3028s_c}{...}
  \item{rea30310_c}{...}
  \item{rea30320_c}{...}
  \item{rea30330_c}{...}
  \item{rea30340_c}{...}
  \item{rea30350_c}{...}
  \item{rea30360_c}{...}
  \item{rea30370_c}{...}
  \item{rea3038s_c}{...}
  \item{rea30410_c}{...}
  \item{rea3042s_c}{...}
  \item{rea30430_c}{...}
  \item{rea30440_c}{...}
  \item{rea30450_c}{...}
  \item{rea30460_c}{...}
  \item{rea30510_c}{...}
  \item{rea3052s_c}{...}
  \item{rea30530_c}{...}
  \item{rea3054s_c}{...}
  \item{rea30550_c}{...}
  \item{rea3_sc1}{simulated reading WLE}
}}
\usage{
xTargetCompetencies_sim
}
\description{
The data set contains simulated dichotomous competence
test items (loosely based on NEPS SC6)
}
\keyword{datasets}
