% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_quality_plots.r
\name{xyplot}
\alias{xyplot}
\title{Plots for multiple imputation}
\usage{
xyplot(pv_obj, var, imputation = "all")
}
\arguments{
\item{pv_obj}{pv_obj by plausible_values() containing at least the sublists
"pv" with the imputed data sets and "indmis", the indicator matrix for
missing variables}

\item{var}{character value denoting the variable against which to plot the
distribution of the imputed variables}

\item{imputation}{indicates which imputation number to plot; defaults to
"all"; otherwise, the imputations to plot are indicated by numeric vectors
(e.g., 1, 2:4, c(1,5))}
}
\value{
ggplot2 objects
}
\description{
The plots show the distribution of the original and imputed data sets per
variable. The plots are analogous to the xyplot, densityplot, and stripplot
of the mice package.
}
