% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimated_pvs.R
\name{xyplot_pvs}
\alias{xyplot_pvs}
\title{Plots for estimated PVs}
\usage{
xyplot_pvs(pv_obj, var, imputation = "all")
}
\arguments{
\item{pv_obj}{pv_obj by plausible_values() containing at least the sublist
"pv" with the imputed data sets}

\item{var}{character value denoting the variable against which to plot the
distribution of the estimated plausible values}

\item{imputation}{indicates which imputation number to plot; defaults to
"all"; otherwise, the imputations to plot are indicated by numeric vectors
(e.g., 1, 2:4, c(1,5))}
}
\value{
ggplot2 objects
}
\description{
The plots show the distribution of the estimated plausible values.
}
