% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_domain}
\alias{get_domain}
\title{Get the domain for which plausible values have been estimated}
\usage{
get_domain(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
the competence domain as a String
}
\description{
Get the domain for which plausible values have been estimated
}
