% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_eap}
\alias{get_eap}
\title{Get the estimated EAP values and standard errors for each person}
\usage{
get_eap(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
a data.frame containing the EAPs and their respective standard errors
}
\description{
Get the estimated EAP values and standard errors for each person
}
