% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_eap_reliability}
\alias{get_eap_reliability}
\title{Get the EAP reliability of the plausible values}
\usage{
get_eap_reliability(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the EAP reliability of the plausible values as a numeric value
}
\description{
Get the EAP reliability of the plausible values
}
