% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_item_difficulties}
\alias{get_item_difficulties}
\title{Get the fixed item difficulties and their standard errors}
\usage{
get_item_difficulties(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
the matrix containing the fixed item difficulties and their estimated
standard errors as numeric values
}
\description{
Get the fixed item difficulties and their standard errors
}
