% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_mean_pv}
\alias{get_mean_pv}
\title{Get the overall mean of the plausible values}
\usage{
get_mean_pv(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the overall mean of the plausible values as a numeric value
}
\description{
Get the overall mean of the plausible values
}
