% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_npv}
\alias{get_npv}
\title{Get the number of plausible values returned by
\code{NEPScaling::plausible_values()}}
\usage{
get_npv(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the number of plausible values as an integer value
}
\description{
Get the number of plausible values returned by
\code{NEPScaling::plausible_values()}
}
