% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_pv_list}
\alias{get_pv_list}
\title{Get the complete list of estimated plausible values and their respective
background data}
\usage{
get_pv_list(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the list of npv plausible values and their respective background data
}
\description{
Get the complete list of estimated plausible values and their respective
background data
}
