% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_rotation}
\alias{get_rotation}
\title{Get info whether test rotation has been considered during estimation}
\usage{
get_rotation(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
whether the test rotation has been considered as a String
}
\description{
Get info whether test rotation has been considered during estimation
}
