% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_starting_cohort}
\alias{get_starting_cohort}
\title{Get the starting cohort for which plausible values have been estimated}
\usage{
get_starting_cohort(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the number of the starting cohort as an integer value (see
https://www.neps-data.de/en-us/datacenter/dataanddocumentation.aspx for their
meaning)
}
\description{
Get the starting cohort for which plausible values have been estimated
}
