% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_type}
\alias{get_type}
\title{Get the type of estimation (longitudinal or cross-sectional)}
\usage{
get_type(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
the type of estimation as a String
}
\description{
Get the type of estimation (longitudinal or cross-sectional)
}
