% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_valid_responses_per_person}
\alias{get_valid_responses_per_person}
\title{Get the number of valid responses (i.e., non-missing) for each test taker}
\usage{
get_valid_responses_per_person(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
a data.frame containing the number of valid responses per person
as an integer value
}
\description{
Get the number of valid responses (i.e., non-missing) for each test taker
}
