% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions_for_pv_obj.R
\name{get_wave}
\alias{get_wave}
\title{Get the wave for which plausible values have been estimated}
\usage{
get_wave(pv_obj)
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}
}
\value{
the number of the wave as an integer value
}
\description{
Get the wave for which plausible values have been estimated
}
