% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values_getter_functions.R
\name{get_wle}
\alias{get_wle}
\title{Get the estimated WLE values and standard errors for each person}
\usage{
get_wle(pv.obj)
}
\arguments{
\item{pv.obj}{return object of function \code{NEPScaling::plausible_values()}}
}
\value{
a data.frame containing the WLEs and their respective standard errors
}
\description{
Get the estimated WLE values and standard errors for each person
}
