% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_values.R
\name{plausible_values}
\alias{plausible_values}
\title{Function for estimating plausible values with NEPS competence data and NEPS
scaled item parameters}
\usage{
plausible_values(SC, domain = c("MA", "RE", "SC", "IC", "LI", "EF", "NR",
  "NT", "OR", "ST", "BA", "CD", "GR"), wave, method = c("ML", "Bayes"),
  path, filetype = c("SPSS", "Stata"), bgdata = NULL, npv = 10L,
  longitudinal = FALSE, rotation = TRUE, nvalid = 3L,
  include.nr = TRUE, control = list(EAP = FALSE, Bayes = list(itermcmc
  = 10000, burnin = 2000, est.alpha = FALSE, thin = 1, tdf = 10, cartctrl1
  = 5, cartctrl2 = 1e-04), ML = list(nmi = 10L, ntheta = 2000,
  normal.approx = FALSE, samp.regr = FALSE, theta.model = FALSE, np.adj =
  8, na.grid = 5, itermcmc = 100, burnin = 50, thin = 1, cartctrl1 = 5,
  cartctrl2 = 1e-04)))
}
\arguments{
\item{SC}{numeric. The starting cohort used for the analysis is
indicated by an integer value (e.g., starting cohort one = 1).}

\item{domain}{character. The competence domain of interest is indicated
by the domain abbreviation as used in the variable names (see Fuß, D., Gnambs,
T., Lockl, K., & Attig, M., 2016).}

\item{wave}{numeric. The wave of competence testing is indicated by
an integer value (e.g., wave one = 1).}

\item{method}{character. Should estimation method be fully Bayesian
("Bayes") or maximum likelihood-bases ("ML")?}

\item{path}{character; file path leading to the location of the
competence data}

\item{filetype}{filetype of competence data (the NEPS RDC provides SPSS
and Stata data files)}

\item{bgdata}{data frame containing background variables. Categorical
variables have to be specified as factors. If \code{bgdata = NULL}, plausible
values are estimated without a background model. Missing data in the covariates
is imputed using sequential classification and regression trees.}

\item{npv}{numeric; number of plausible values to be estimated;
defaults to 10.}

\item{longitudinal}{logical. TRUE indicating that a multidimensional model with
the measurement points as its dimensions is to be estimated. Defaults to FALSE.}

\item{rotation}{logical. TRUE indicating that the competence scores
should be corrected for the rotation design of the booklets. Defaults to
TRUE. If both longitudinal and rotation are TRUE, rotation is set to FALSE
automatically.}

\item{nvalid}{numeric; minimum number of responses test takers had to
give for inclusion into the estimation process. NEPS default is at least three.}

\item{include.nr}{logical; whether the number of not-reached items as a proxy
for processing speed should be included in the background model (the default
is TRUE)}

\item{control}{list of additional options. If \code{EAP = TRUE}, the EAPs
will be returned as well. Furthermore, additional control options for "ML"
(again in form of a list) are \code{nmi} (the number of multiple imputations
if there is missing data in the covariates) and arguments concerning the ML
estimation given to TAM::tam.pv() (therefore, see TAM for further control
options). Analogously, control options for "Bayes" (in form of a list) are
"itermcmc" (number of MCMC iterations), "burnin" (number of burn-in iterations),
"est.alpha" (whether item discriminations of the graded response model are to
be fixed or estimated freely; default: fixed), "thin" (thinning interval in
MCMC), "tdf" (df of multiv. t proposal distr. for category cut-offs for
ordinal items), "cartctrl1" (min. number of observations in any terminal CART
node) and "cartctrl2" (min. decrease of overall lack of fit by each CART
split). NOTE that "Bayes" control options are also applied to the CART
algorithm for "ML" estimation.}
}
\value{
\code{plausible_values()} returns an object of class \code{pv.obj}
containing:
\describe{
\item{SC}{Starting cohort that plausible values were estimated for}
\item{domain}{Competence domain that plausible values were estimated for}
\item{wave}{Wave that plausible values were estimated for}
\item{method}{Estimation algorithm that was used for estimating plausible
values (\code{ML} for the EM algorithm implemented in \code{TAM} or \code{Bayes}
for the Gibbs Sampler implemented in \code{LaRA})}
\item{type}{Whether cross-sectional ("cross") or longitudinal ("long")
plausible values were estimated}
\item{rotation}{In most assessments the position of the competence test was
rotated with another competence domain. If this was the case for the
specific estimation, this variable indicates whether the correction was
applied. Depending on the estimation context, this variable may have been
automatically set by the function and thus differ from user input}
\item{nvalid}{The minimum number of answers a test taker must have given}
\item{model}{If the method \code{ML} was chosen, a Rasch or a Partial Credit
Model was estimated, depending on the item format. If \code{Bayes} was chosen,
a Graded Response Model was estimated.}
\item{n.valid}{A \code{data.frame} containing the \code{ID_t} and the
number of valid responses given by the respective individual}
\item{npv}{The number of plausible values that are returned by the function}
\item{control}{The control variables that were applied to fine-tune the
estimation algorithms}
\item{position}{A \code{data.frame} containing the \code{ID_t} and the
position the respective individual got the testlet in (first or second)}
\item{abs.correction}{The total amount by which the scale was shifted in
longitudinal estimation (consisting of linking constants and dropout
corrections)}
\item{variance.PV}{The overall variance of all persons' abilities}
\item{mean.PV}{The overall mean of all persons' abilities}
\item{pv}{A list of \code{data.frame}s containing one plausible value each
and the imputed data set that was used to estimate the plausible value.
Additionally, if \code{include.nr} was specified, the background model is
enriched by the number of not reached items (\code{nr}) per test taker as a
proxy for response times.}
\item{eap}{A \code{data.frame} containing the \code{ID_t} and the ability
EAP value for the respective individual}
\item{regr.coeff}{The regression coefficients of the latent regression of
the ability}
\item{alpha}{If the method \code{Bayes} and, in the controls, \code{est.alpha}
were chosen, the item discrimination parameters (as the EAPs) are returned}
\item{EAP.rel}{For method \code{ML} the EAP reliability is returned}
}
}
\description{
Function for estimating plausible values with NEPS competence data and NEPS
scaled item parameters
}
\note{
For an extensive example, see the accompanying NEPS Survey Paper.
The function will only work with NEPS data. To access NEPS data see
https://www.neps-data.de/en-us/datacenter/dataaccess.aspx.
}
\examples{
\dontrun{
rm(list = ls())
library(NEPScaling)
library(foreign)
## read in data object for conditioning variables
data(bg_data)
bg_data$gender <- as.factor(bg_data$gender)
## define path to NEPS competence data
path <- 'Usr/NEPS-data/'
setwd(path)
## save simulated target competencies to path
data(xTargetCompetencies_sim)
write.dta(dataframe = xTargetCompetencies_sim, file = 'SC6_xTargetCompetencies_sim.dta')
## estimate default number of 10 plausible values
## note: the example background data is completely observed!
result <- plausible_values(SC = 6
                           , domain = 'RE'
                           , wave = 3
                           , bgdata = bg_data
                           , path = path
                           , method = 'Bayes'
                           , filetype = 'Stata'
                           )
}

}
\references{
Albert, J. H. (1992). Bayesian estimation of normal ogive item
response curves using Gibbs sampling. \emph{Journal of Educational
Statistics}, \emph{17}(3), 251-269.

Azur, M. J., Stuart, E. A., Frangakis, C., & Leaf, P. J. (2011).
Multiple imputation by chained equations: what is it and how does it work?
\emph{International Journal of Methods in Psychiatric Research, 20}(1), 40-49.

Blossfeld, H.-P., Rossbach, H.-G., & von Maurice,  J. (Eds.) (2011).
Education as a Lifelong Process - The German National Educational Panel Study
(NEPS). \emph{Zeitschrift fuer Erziehungswissenschaft, Sonderheft 14}.

Burgette, L. F., & Reiter, J. P. (2010). Multiple imputation for
missing data via sequential regression trees. \emph{American Journal of
Epidemiology}, \emph{172}(9), 1070-1076.

Mislevy, R. J. (1991). Randomization-based inference about latent
variables from complex samples. \emph{Psychometrika, 56}(2), 177-196.

Pohl, S., & Carstensen, C. H. (2012). NEPS technical report -
Scaling the data of the competence tests.

Scharl, A., Carstensen, C. H., & Gnambs, T. (2019).
\emph{Estimating Plausible Values with NEPS Data: An Example Using Reading
Competence in Starting Cohort 6 (NEPS Survey Paper No. XX)}. Bamberg: Leibniz
Institute for Educational Trajectories, National Educational Panel Study.

Tanner, M. A., & Wong, W. H. (1987). The calculation of posterior
distributions by data augmentation. \emph{Journal of the American Statistical
Association}, \emph{82}(398), 528-549.

Weirich, S., Haag, N., Hecht, M., Boehme, K., Siegle, T., & Luedtke,
O. (2014). Nested multiple imputation in large-scale assessments.
\emph{Large-Scale Assessments in Education, 2}(1), 9.
}
