% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pv.R
\name{write_pv}
\alias{write_pv}
\title{Function to write pv_obj to file}
\usage{
write_pv(pv_obj, path, ext = c("SPSS", "Stata", "Mplus"))
}
\arguments{
\item{pv_obj}{return object of function \code{NEPSscaling::plausible_values()}}

\item{path}{path to the directory to save the plausible values to}

\item{ext}{format (either "SPSS", "Stata" (version 14) or "Mplus" compatible)}
}
\value{
A data set per imputed plausible value in the specified format and
for Mplus a content and variable names file.
}
\description{
Function to write pv_obj to file
}
\examples{
\dontrun{
rm(list = ls())
library(NEPSscaling)
# the pv_obj results was created by plausible_values()
path <- "chosen/directory/"
write_pv(results, path, "Mplus")
write_pv(results, path, "Stata")
write_pv(results, path, "SPSS")
}

}
